+++ ChangeLog.md
@@ -1,7 +1,7 @@
 # Write Your Python Program - CHANGELOG
 
 0.8.1 (2020-11-24)
-  * Fix command execution under powershell
+  * Fix command execution under windows and powershell
   * Allow `List[T]` if `T` is an `Enum`, `Mixed` or `DefinedLater`
   * Better formatting of tracebacks
 0.8.0 (2020-11-22)